requirements.txt
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
numpy
Pillow
pysesame3
zmq
line-bot-sdk